requirements.txt
beautifulsoup4
chat-exporter
discord.py
requests
better-profanity